GameManager en cours de programmation -> Il me faudra d'autres scènes afin notemment de changer la difficulté des questions dans le dictionnairedu game manager

+++ Assets/Scripts/NumberData.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 public class NumberData : ScriptableObject
 {
     public string numberName;
-    public float buildTime;
+    public int buildTime;
     public GameObject prefab;
 }
